[stable30] feat: Use new interface to manage metadata

+++ lib/Versions/AbstractS3VersionBackend.php
@@ -24,20 +24,24 @@ declare(strict_types=1);
 namespace OCA\FilesVersionsS3\Versions;
 
 use OC\Files\ObjectStore\S3ConnectionTrait;
+use OCA\DAV\Connector\Sabre\Exception\Forbidden;
 use OCA\Files_Versions\Versions\IDeletableVersionBackend;
-use OCA\Files_Versions\Versions\INameableVersionBackend;
+use OCA\Files_Versions\Versions\IMetadataVersionBackend;
 use OCA\Files_Versions\Versions\IVersion;
 use OCA\Files_Versions\Versions\IVersionBackend;
 use OCP\Files\File;
 use OCP\Files\FileInfo;
+use OCP\Files\Node;
+use OCP\Files\NotFoundException;
 use OCP\Files\Storage\IStorage;
 use OCP\IUser;
+use OCP\IUserSession;
 
-abstract class AbstractS3VersionBackend implements IVersionBackend, INameableVersionBackend, IDeletableVersionBackend {
-	protected $versionProvider;
-
-	public function __construct(S3VersionProvider $versionProvider) {
-		$this->versionProvider = $versionProvider;
+abstract class AbstractS3VersionBackend implements IVersionBackend, IMetadataVersionBackend, IDeletableVersionBackend {
+	public function __construct(
+		private S3VersionProvider $versionProvider,
+		private IUserSession $userSession,
+	) {
 	}
 
 	abstract public function useBackendForStorage(IStorage $storage): bool;
@@ -66,6 +70,10 @@ abstract class AbstractS3VersionBackend implements IVersionBackend, INameableVer
 	}
 
 	public function rollback(IVersion $version) {
+		if (!$this->currentUserHasPermissions($version->getSourceFile(), \OCP\Constants::PERMISSION_UPDATE)) {
+			throw new Forbidden('You cannot restore this version because you do not have update permissions on the source file.');
+		}
+
 		$source = $version->getSourceFile();
 		$s3 = $this->getS3($source);
 		if ($s3) {
@@ -98,19 +106,39 @@ abstract class AbstractS3VersionBackend implements IVersionBackend, INameableVer
 		throw new \Exception("Requested s3 version for a file not stored in s3");
 	}
 
-	public function setVersionLabel(IVersion $version, string $label): void {
-		$source = $version->getSourceFile();
-		$s3 = $this->getS3($source);
-		if ($s3) {
-			$this->versionProvider->setVersionLabel($s3, $this->getUrn($version->getSourceFile()), $version->getRevisionId(), $label);
-		}
-	}
-
 	public function deleteVersion(IVersion $version): void {
+		if (!$this->currentUserHasPermissions($version->getSourceFile(), \OCP\Constants::PERMISSION_DELETE)) {
+			throw new Forbidden('You cannot delete this version because you do not have delete permissions on the source file.');
+		}
+
 		$source = $version->getSourceFile();
 		$s3 = $this->getS3($source);
 		if ($s3) {
 			$this->versionProvider->deleteVersion($s3, $this->getUrn($version->getSourceFile()), $version->getRevisionId());
 		}
+	}
+
+	public function setMetadataValue(Node $node, int $revision, string $key, string $value): void {
+		if (!$this->currentUserHasPermissions($node, \OCP\Constants::PERMISSION_UPDATE)) {
+			throw new Forbidden('You cannot update the version\'s metadata because you do not have update permissions on the source file.');
+		}
+
+		$versions = $this->getVersionsForFile($this->userSession->getUser(), $node);
+		$version = array_values(array_filter($versions, fn (IVersion $version) => $version->getTimestamp() === $revision))[0] ?? null;
+
+		$s3 = $this->getS3($node);
+		if ($s3 && $version) {
+			$this->versionProvider->setVersionMetadata($s3, $this->getUrn($node), $version->getRevisionId(), $key, $value);
+		}
+	}
+
+	private function currentUserHasPermissions(FileInfo $sourceFile, int $permissions): bool {
+		$currentUserId = $this->userSession->getUser()?->getUID();
+
+		if ($currentUserId === null) {
+			throw new NotFoundException("No user logged in");
+		}
+
+		return ($sourceFile->getPermissions() & $permissions) === $permissions;
 	}
 }
